Clean up service listener
